Final checks for code release

+++ test/TestOMPL3DArticulated.cpp
@@ -1,3 +1,5 @@
+/** \authors Sipu Ruan, Karen L. Poblete */
+
 #include "hrm/config.h"
 #include "hrm/planners/ompl_interface/OMPL3DArticulated.h"
 #include "hrm/test/util/GTestUtils.h"
@@ -65,7 +67,7 @@ void TestOMPLPlanner(const int plannerIdx, const int samplerIdx) {
     res.solved = omplPlanner.isSolved();
     res.graphStructure.edge = omplPlanner.getEdges();
     res.graphStructure.vertex = omplPlanner.getVertices();
-    res.solutionPath.cost = omplPlanner.getPathLength();
+    res.solutionPath.cost = static_cast<double>(omplPlanner.getPathLength());
     res.solutionPath.solvedPath = omplPlanner.getSolutionPath();
     res.planningTime.totalTime = omplPlanner.getPlanningTime();
